+++ tests/Model/YamlIndent/YamlIndentCheckerTest.php
@@ -6,6 +6,7 @@ namespace YamlStandards\Model\YamlIndent;
 
 use PHPUnit\Framework\TestCase;
 use YamlStandards\Model\Config\StandardParametersData;
+use YamlStandards\Model\Config\YamlStandardConfigDefinition;
 use YamlStandards\Result\Result;
 
 class YamlIndentCheckerTest extends TestCase
@@ -75,10 +76,10 @@ class YamlIndentCheckerTest extends TestCase
     }
 
     /**
-     * @return \YamlStandards\Model\Config\StandardParametersData|\PHPUnit_Framework_MockObject_MockObject
+     * @return \YamlStandards\Model\Config\StandardParametersData
      */
-    private function getStandardsParametersData()
+    private function getStandardsParametersData(): StandardParametersData
     {
-        return new StandardParametersData(4, 4, 4, null);
+        return new StandardParametersData(4, 4, 4, YamlStandardConfigDefinition::CONFIG_PARAMETERS_SERVICE_ALIASING_TYPE_VALUE_SHORT);
     }
 }