+++ tests/Model/YamlServiceAliasing/YamlServiceAliasingFixerTest.php
@@ -1,6 +1,6 @@
 <?php
 
-declare(strict_types = 1);
+declare(strict_types=1);
 
 namespace YamlStandards\Model\YamlServiceAliasing;
 
@@ -72,6 +72,6 @@ class YamlServiceAliasingFixerTest extends TestCase
      */
     private function getStandardsParametersData(string $type): StandardParametersData
     {
-        return new StandardParametersData(null, null, null, $type);
+        return new StandardParametersData(4, 4, 2, $type);
     }
 }