+++ tests/Model/YamlSpacesBetweenGroups/YamlSpacesBetweenGroupsFixerTest.php
@@ -6,6 +6,7 @@ namespace YamlStandards\Model\YamlSpacesBetweenGroups;
 
 use PHPUnit\Framework\TestCase;
 use YamlStandards\Model\Config\StandardParametersData;
+use YamlStandards\Model\Config\YamlStandardConfigDefinition;
 
 class YamlSpacesBetweenGroupsFixerTest extends TestCase
 {
@@ -50,10 +51,10 @@ class YamlSpacesBetweenGroupsFixerTest extends TestCase
 
     /**
      * @param int $level
-     * @return \YamlStandards\Model\Config\StandardParametersData|\PHPUnit_Framework_MockObject_MockObject
+     * @return \YamlStandards\Model\Config\StandardParametersData
      */
-    private function getStandardsParametersData(int $level)
+    private function getStandardsParametersData(int $level): StandardParametersData
     {
-        return new StandardParametersData(4, 4, $level, null);
+        return new StandardParametersData(4, 4, $level, YamlStandardConfigDefinition::CONFIG_PARAMETERS_SERVICE_ALIASING_TYPE_VALUE_SHORT);
     }
 }